Added logic to handle a bug on promotion/demotion.
Moved pl.writeMMRdiff(int64, int64, ...int) down in mainAPI(*player2).
Added sentinel error ErrPromoted for when a player is promoted/demoted.

Modified getLadder(*http.Client, string) (int64, error) to return
an error if the player is promoted or demoted.

Changed (*player) run(user) so that if using useAPI = true, and there's
a BattlenetError, or promotion, it simply grabs MMR from the last file.

Added a real example of promotion file w/ the old ladderID of 301869:
  api/ladder/json/us-Hengf-promoted-G1toP3-301869.json